Change id from int to long to increase total number of available short urls.

+++ src/main/java/com/example/urlshortener/service/CounterService.java
@@ -1,5 +1,5 @@
 package com.example.urlshortener.service;
 
 public interface CounterService {
-    int getCounter();
+    long getCounter();
 }
